fix: reverts `vite-tsconfig-paths`, fixing path resolution in build
`vite-tsconfig-paths` helps vite to resolve "absolute" paths when compiling.

This plugin doesn't play nicely with `vite-plugin-dts`.

This is a known problem, with GH issues in both repos.

Another plugin `vite-plugin-lib`, which is based on `vite-plugin-dts`, apparently handles all of this, but I couldn't get it to work.

WIth `vite-tsconfig-paths` being removed, we also have no need for `eslint-plugin-path`, because we are using relative paths, so it's also removed.

Tweaked `tsconfig.json` and `vite.config.ts` in response to these changes.

The packages now build correctly and the types work.