fix(tooltip): set default openDelay to 500ms
In 76a6ea858fffcf87e784d9f7aef3727723b40c81 all `onClick` events were changed to `onPointerUp`. This allows Apple Pencil to work on buttons with tooltips.

Unfortunately, this changes click behaviour in a subtle way. With `onClick`, if you mouse down on an element, then move the mouse to a different element before releasing the mouse, nothing happens. With `onPointerUp`, the different element is clicked.

An alternate solution to is set a default `openDelay` on Tooltips to a higher value (default is no delay). I did a bit of searching and found it's actually considered a best-practice to make the delay ~500ms by default, and only reduce the delay in specific use-cases where it makes sense.

Tooltip now has a default `openDelay` of 500ms.